Disable Weaponinfo and Vehoff limits on PreCE
Because not sure about "ReplaceXrefs" offsets (VehOff causes a bug when shooting a driver/passenger and then entering the vehicle from their side. The corpse to still be there despite animation of throwing them out  playing. The corpse will perform getting hurt animation and player will just enter into them. The player won't be in control of the car and won't be able to get out)